Add item action menu, product note modal, and catalog barcode linking
- Add ItemActionMenu bottom sheet: click on item opens menu with
  "Modifica" and "Modifica nota articolo" options
- Add ProductNoteModal: simple editor for product notes, saves directly
  to catalog via barcode
- Remove "Commento Prodotto" field from ItemDetailModal (now handled
  by dedicated modal)
- Show product notes in autocomplete suggestions (backend + frontend)
- Link catalog barcode when selecting a product from autocomplete, so
  EAN is pre-filled when checking items off
- Show product notes in LoadVerification item rows
- Add inline note editing with save button in ShoppingListForm

+++ backend/app/models/shopping_list.py
@@ -141,7 +141,7 @@ class ShoppingListItem(BaseModel):
     grocy_product_name = Column(String(255), nullable=True)
 
     # Quantity (optional) - can be integer for pieces or float for weight
-    quantity = Column(Integer, default=1, nullable=False)  # Store as int, frontend handles decimals for weight
+    quantity = Column(Float, default=1, nullable=False)  # Supports decimals for weight-based items (e.g. 0.5 kg)
     unit = Column(String(50), nullable=True)  # 'pz' for pieces, 'kg' for weight
 
     # Checked/purchased status
